implement field components translation (unit, attack, move ... etc), break moving since myField property is removed from Field class (path[0] should be used instead but in previous implementation path[0] was the first field in front of the unit). Placin units in initialization message is broken, config.defaultPositions.Conains(point) does not work for some reason. Ready for merge

+++ Source/client-view/src/main/java/view/ResourceManager.java
@@ -216,9 +216,9 @@ public class ResourceManager {
 		return ResourceManager.ASSETS_TYPE;
 	}
 
-	public Image getTerrain(String terrainName, int intensity) {
+	public Image getTerrain(String terrainName, float intensity) {
 
-		return this.assets.get(this.constructTerrainKey(terrainName, intensity));
+		return this.assets.get(this.constructTerrainKey(terrainName, (int) intensity));
 
 	}
