+++ Source/root/src/main/java/root/model/component/Terrain.java
@@ -2,16 +2,16 @@ package root.model.component;
 
 public class Terrain {
 
-	// these are lowercase beacuse they are mapped to the sampe-named res files
+	// these are lowercase because they are mapped to the same-named resource files
 	public static enum TerrainType {
 		mountains,
 		water
 	}
 
 	private Terrain.TerrainType type;
-	private int intensity;
+	private float intensity;
 
-	public Terrain(Terrain.TerrainType type, int intensity) {
+	public Terrain(Terrain.TerrainType type, float intensity) {
 		this.type = type;
 		this.intensity = intensity;
 	}
@@ -20,7 +20,7 @@ public class Terrain {
 		return this.type;
 	}
 
-	public int getIntensity() {
+	public float getIntensity() {
 		return this.intensity;
 	}
